Whisper-Llama-Project/requirements.txt
streamlit
whisper
openai-whisper
pydub
ffmpeg-python
groq
pyngrok
python-dotenv
